refactor(tag-input): adjust file directory
